~ moved calculations to separate lib file
